Move command into "/config announce"
